imguizmo integration

+++ handle_vec.hpp
@@ -19,9 +19,9 @@ template <typename T, typename Storage = uint32_t> class HandleVector {
     constexpr auto end(this auto&& self) { return self.storage.end(); }
     constexpr auto& front() { return storage.front(); }
     constexpr auto size() const { return storage.size(); }
-    constexpr auto& at(size_t idx) { return storage.at(idx); }
+    constexpr auto& at(u64 idx) { return storage.at(idx); }
     constexpr auto& at(Handle handle) { return at(find_idx(handle)); }
-    constexpr auto& operator[](size_t idx) { return storage.at(idx); }
+    constexpr auto& operator[](u64 idx) { return storage.at(idx); }
 
     constexpr void insert(Handle h, auto&& t) {
         const auto it = find_insertion_it(h);
@@ -53,8 +53,8 @@ template <typename T, typename Storage = uint32_t> class HandleVector {
         if(it == handles.end()) { return nullptr; }
         return &at(std::distance(handles.begin(), it));
     }
-    constexpr size_t find_idx(Handle handle) const { return std::distance(handles.begin(), find_handle_it(handle)); }
-    constexpr Handle handle_at(size_t idx) const { return handles.at(idx); }
+    constexpr u64 find_idx(Handle handle) const { return std::distance(handles.begin(), find_handle_it(handle)); }
+    constexpr Handle handle_at(u64 idx) const { return handles.at(idx); }
 
   private:
     constexpr auto find_handle_it(this auto&& self, Handle h) {
